+++ package.json
@@ -2,9 +2,10 @@
   "name": "todolistreview",
   "version": "1.0.0",
   "description": "![](https://img.shields.io/badge/Microverse-blueviolet)",
-  "main": "index.js",
+  "private": true,
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test2": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
+    "test": "jest",
     "start": "webpack serve --open",
     "build": "webpack --config webpack.config.js"
   },
@@ -20,6 +21,9 @@
   },
   "homepage": "https://github.com/aron-helu/todoListReview#readme",
   "devDependencies": {
+    "@babel/core": "^7.18.9",
+    "@babel/plugin-transform-modules-commonjs": "^7.18.6",
+    "@babel/preset-env": "^7.18.9",
     "babel-eslint": "^10.1.0",
     "css-loader": "^6.7.1",
     "eslint": "^7.32.0",
@@ -30,5 +34,13 @@
     "webpack": "^5.73.0",
     "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.9.3"
+  },
+  "dependencies": {
+    "babel-jest": "^28.1.3",
+    "electron-webpack-boilerplate": "^1.0.5",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
+    "lodash": "^4.17.21",
+    "ts-jest": "^28.0.7"
   }
 }
