+++ src/index.js
@@ -1,16 +1,10 @@
-import './index.css';
-import * as Store from './modules/store.js';
-import UIDisplay from './modules/showTask.js';
-import taskArr from './modules/toDo.js';
-import removeTask from './modules/remove.js';
-
-class Task {
-  constructor(description, completed, index) {
-    this.description = description;
-    this.completed = completed;
-    this.index = index;
-  }
-}
+import './style.css';
+import * as Store from './modules/local-storage.js';
+import UIDisplay from './modules/userInterface.js';
+import taskArr from './modules/taskarr.js';
+import removeTask from './modules/removeTask.js';
+import addNewTask from './modules/addtask.js';
+import filterAll from './modules/clearAll.js';
 
 const syncIcon = document.getElementById('sync');
 
@@ -33,13 +27,8 @@ const addIcon = document.getElementById('add-icon');
 
 addInput.addEventListener('keypress', (e) => {
   if (e.key === 'Enter' && addInput.value) {
-    const task = new Task(addInput.value, false, taskArr.length + 1);
-    taskArr.push(task);
-    // Prevent submit
+    addNewTask(addInput.value);
     e.preventDefault();
-    UIDisplay.createTask(task);
-    Store.addTask(taskArr);
-
     addInput.value = '';
   } else {
     addInput.setAttribute('required', '');
@@ -47,16 +36,17 @@ addInput.addEventListener('keypress', (e) => {
 });
 
 addIcon.addEventListener('click', (e) => {
-  if (addInput.value) {
-    const task = new Task(addInput.value, false, taskArr.length + 1);
-    taskArr.push(task);
-    // Prevent submit
+  if (e.key === 'Enter' && addInput.value) {
+    addNewTask(addInput.value);
     e.preventDefault();
-    UIDisplay.createTask(task);
-    Store.addTask(taskArr);
-
     addInput.value = '';
   } else {
     addInput.setAttribute('required', '');
   }
+});
+
+const clearAll = document.getElementById('clear-all');
+
+clearAll.addEventListener('click', () => {
+  filterAll(taskArr);
 });
