+++ src/style.css
@@ -9,9 +9,11 @@
 
 body {
   font-family: 'Space Grotesk', sans-serif;
+  background: url('./bg.jpg');
   background-size: cover;
   background-position: center;
   background-repeat: no-repeat;
+  background-color: rgba(243, 173, 75, 0.845);
   background-blend-mode: soft-light;
   display: flex;
   flex-direction: column;
@@ -22,13 +24,31 @@ body {
   height: 100vh;
 }
 
-#todo-list {
+#todo-list,
+#hero {
   display: flex;
   flex-direction: column;
   width: 80%;
   background-color: rgb(247, 247, 248);
-  border: 10px solid whitesmoke;
-  border-image: url('./background-with-floral-elements-border-free-vector.jpg') 20 stretch;
+  box-shadow: 5px 10px 8px rgba(202, 67, 29, 0.492);
+}
+
+#todo-list {
+  max-height: 80%;
+}
+
+#hero {
+  background: none;
+  box-shadow: none;
+  align-items: center;
+  gap: 1rem;
+}
+
+#hero h1 {
+  font-family: 'Darker Grotesque', sans-serif;
+  color: rgb(219, 45, 67);
+  font-size: 3rem;
+  font-weight: 700;
 }
 
 #todo-list a {
@@ -36,6 +56,10 @@ body {
   padding: 1rem 1.3rem;
   text-align: center;
   text-decoration: none;
+  background-image: linear-gradient(90deg, rgb(168, 2, 111), rgb(219, 45, 67), rgb(231, 104, 56), rgb(255, 153, 0));
+  background-clip: text;
+  -webkit-background-clip: text;
+  color: transparent;
   font-weight: 600;
 }
 
@@ -72,7 +96,7 @@ li {
 
 #header h2 {
   font-family: 'Darker Grotesque', sans-serif;
-  background-image: linear-gradient(90deg, rgb(78, 73, 76), rgb(75, 67, 68), rgb(219, 174, 157));
+  background-image: linear-gradient(90deg, rgb(168, 2, 111), rgb(219, 45, 67), rgb(231, 104, 56), rgb(255, 153, 0));
   background-clip: border-box;
   -webkit-background-clip: text;
   color: transparent;
@@ -117,6 +141,9 @@ i[class^="fa"] {
 
 #task-list {
   width: 100%;
+  overflow: hidden;
+  overflow-y: scroll;
+  max-height: 350px;
 }
 
 .task {
